# Replace the distance search of finding hydrogens with a dictionary lookup