Changes made in this Pull Request:

In the original water bridge analysis implementation, the hydrogen associated with the hydrogen bond donor heavy atom were looked up through a distance search, which is repeated every single frame. This PR attempts to generate the dictionary of hydrogen bond donor heavy atom vs hydrogen pair in the first frame and do a dictionary lookup for the rest of the frames.

This PR should reduce the processing time for all the rest of the frames at the cost of increasing the time for the first frame. This PR should also reduce the computational overhead for turning on the option of `update_selection `.

Here are the benchmark results.

| Frames | new update_selection=False (sec) | old update_selection=False (sec) | new update_selection=True (sec) | old update_selection=True (sec) |
|--------|----------------------------------|----------------------------------|---------------------------------|---------------------------------|
| 1      | 12.0                             | 1.9                              | 10.0                            | 2.5                             |
| 10     | 13.7                             | 7.8                              | 14.0                            | 13.4                            |
| 100    | 37.9                             | 63.3                             | 59.7                            | 119.0                           |
| 1000   | 288.6                            | 649.6                            | 478.7                           | 1137.1                          |

Combined with #2480, the water bridge analysis is now 4X faster than the original implementation and I felt there is probably little room for performance improvement now.

PR Checklist
------------
 - [ ] Tests?
 - [ ] Docs?
 - [ ] CHANGELOG updated?
 - [ ] Issue raised/referenced?